Renamed solution
I'm making this a bit more generic. I want to have this as a general
purpose core for a gameboy emulator with scope to add other systems with
new config & hardware implementations. This will essentially be a common
back end for analytics based emulation (e.g. optimization) and real-time
stuff.